[Refactor] 로직 리펙토링
