+++ BookBridge/BookBridge/Utility/Extension/UIImageExtension.swift
@@ -1,0 +1,21 @@
+//
+//  UIImageExtension.swift
+//  BookBridge
+//
+//  Created by 노주영 on 2/6/24.
+//
+
+import Foundation
+
+extension UIImage {
+    func aspectFittedToHeight(_ newHeight: CGFloat) -> UIImage {
+        let scale = newHeight / self.size.height
+        let newWidth = self.size.width * scale
+        let newSize = CGSize(width: newWidth, height: newHeight)
+        let renderer = UIGraphicsImageRenderer(size: newSize)
+        
+        return renderer.image { _ in
+            self.draw(in: CGRect(origin: .zero, size: newSize))
+        }
+    }
+}
